update jsvar editor code to UI controlled
