+++ SubathonManager.UI/Converters/Converters.cs
@@ -228,4 +228,12 @@ namespace SubathonManager.UI.Converters
         public object ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
             => throw new NotImplementedException();
     }
+    
+    public class StringToBoolConverter : IValueConverter
+    {
+        public object Convert(object? value, Type targetType, object? parameter, CultureInfo culture)
+            => bool.TryParse(value as string, out var b) && b;
+        public object ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture)
+            => (value is bool ? value.ToString() : "False") ?? "False";
+    }
 }
